Waiting for tetsting

+++ Enteties/Transfer.go
@@ -1,9 +1,9 @@
 package Enteties
 
 type Transfer struct {
-	FromAccountID int `json:"fromAccountID"`
-	ToAccountID   int `json:"toAccountID"`
-	FromStudentID int `json:"fromStudentID"`
-	ToStudentID   int `json:"toStudentID"`
-	Amount        int `json:"Amount"`
+	FromAccountID int     `json:"fromAccountID"`
+	ToAccountID   int     `json:"toAccountID"`
+	FromStudentID int     `json:"fromStudentID"`
+	ToStudentID   int     `json:"toStudentID"`
+	Amount        float64 `json:"Amount"`
 }
